refactor: rename `context/newDataExplorer` to `context/fluxQueryBuilder`
