+++ src/dataExplorer/components/MeasurementSelector.tsx
@@ -6,7 +6,7 @@ import SelectorTitle from 'src/dataExplorer/components/SelectorTitle'
 import SearchableDropdown from 'src/shared/components/SearchableDropdown'
 
 // Context
-import {NewDataExplorerContext} from 'src/dataExplorer/context/newDataExplorer'
+import {NewDataExplorerContext} from 'src/dataExplorer/context/fluxQueryBuilder'
 import {MeasurementsContext} from 'src/dataExplorer/context/measurements'
 
 // Types
